Move tx bins abstraction to a new module
